Add altcoin support Add infrastructure to enable configuring the prefix (byte or string) used by various addresses formats dependant on which blockchain is in use.
This changes the public API of `rust-bitcoin` quite significantly for
`Address`.

Add support for dogecoin and litecoin

Notes:

- Segwit is not implemented for all Litecoin networks (only mainnet
and testnet) because I was unable to find out the segwit prefix for
signet (is this a thing for litecoin?) or regtest.

- Dogecoin does not currently support segwit, this is mirrored by this
patch.

Ported from: https://github.com/tcharding/rust-bitcoin/compare/altcoin-support-v0.26.2