Do not fail serialization with excess data
Deserialization is failing for the block headers coming from `dogecoind`,
I believe this is because dogecoin includes an optional `auxpow` pointer
in the block header used during merge mining. When this data is present
we fail deserilazation because of the explicitly check that
all data was consumed.

Inhibit the check that asserts all data was consumed during
deserialization.

Ref: 4dab8f9dbe69eea870b03e4ea4300b6415639cc2